+++ src/components/HeroesItem/HeroesItem.tsx
@@ -13,18 +13,18 @@ export const HeroesItem = (props: IProps) => {
   return (
     <div className="heroes">
 
-          <div className="heroes__avatar">
-            <img src={`${hero.thumbnail.path}.${hero.thumbnail.extension}`} alt="avatar" />
-          </div>
+      <div className="heroes__avatar">
+          <img src={`${hero.thumbnail.path}.${hero.thumbnail.extension}`} alt="avatar" />
+      </div>
 
-          <div className="heroes__info">
-            <span>{hero.name}</span>
-          </div>
+      <div className="heroes__info">
+        <span>{hero.name}</span>
+      </div>
 
-          <div className="heroes__btn">
-            <Link to={`/comics/${hero.id}`}>See more bin</Link>
-          </div>
+      <div className="heroes__btn">
+        <Link to={`/comics/${hero.id}`}>See more bin</Link>
+      </div>
 
-          </div>
+    </div>
   )
 }
