+++ src/components/Home/Home.tsx
@@ -1,17 +1,15 @@
 import React, {Component} from "react";
-import { Heroes } from '../heroes'
+import { Heroes } from '../Heroes'
 import './home.scss'
 
 export class Home extends Component {
-
   render () {
     return (
-      <React.Fragment>
+      <>
           <main className="main">
             <Heroes />
           </main>
-      </React.Fragment>
-
+      </>
     )
   }
 }
