+++ src/components/PageError/PageError.tsx
@@ -1,13 +1,21 @@
 import React from "react";
 import './pageError.scss'
 import error from '../../img/error.jpg'
+import { withRouter } from "react-router-dom"
 
-export const PageError = () => {
+const PageError = (props: any) => {
+  const errorPath = <span className="span-path">{props.location.pathname}</span>
+  console.log(props)
   return (
     <div className="container">
       <div className="img-wrapper">
         <img src={error} alt="404" />
+        <div className="error-path">
+          Page not found: {errorPath}
+        </div>
       </div>
     </div>
   )
 }
+
+export default withRouter(PageError)
