+++ src/components/PageError/index.ts
@@ -1,1 +1,2 @@
-export {PageError} from './PageError'
+import PageError from './PageError'
+export { PageError }
