+++ src/components/SearchPanel/index.ts
@@ -1,0 +1,1 @@
+export {SearchPanel} from './SearchPanel'
