+++ src/components/SearchPanel/search.scss
@@ -28,6 +28,6 @@
   }
 }
 
-.css-nz481w-MuiInputBase-input-MuiInput-input {
-  background-color: #c62828;
+.custom-input {
+  color: red;
 }
