+++ src/components/app/App.tsx
@@ -1,21 +1,16 @@
 import React from "react"
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom"
-import { Header } from '../header'
+import { Header } from '../Header'
 import { Home } from '../Home'
 import { PageError } from "../PageError"
 import { ComicsPage } from '../ComicsPage'
-import { Box } from '@mui/material'
 import './app.scss'
-
-
-
 
 
 export class App extends React.Component {
   render() {
     return (
         <>
-        
           <Router>
             <Header/>
             <Switch>
@@ -25,7 +20,6 @@ export class App extends React.Component {
             </Switch>
           </Router>
 </>
-        
     )
   }
 }
