+++ src/components/header/index.ts
@@ -1,1 +1,1 @@
-export {Header} from './header'
+export {Header} from './Header'
