--- src/components/search-panel/index.ts
@@ -1,1 +1,0 @@
-export {SearchPanel} from './Search'
